refactor(sources): split discovery/mod.rs into sln.rs and plcproj.rs
Addresses review: discovery/mod.rs had grown past the project's 1000-line
module limit. Splits out .sln -> .tsproj -> PrjFilePath resolution
(sln.rs) and .plcproj parsing/merging (plcproj.rs), each with their
directly-related tests. mod.rs now holds just the top-level dispatch
(discover/detect_beremiz/detect_twincat/detect_fallback) at 952 lines.

No behavior change -- pure move, all 39 discovery tests pass unchanged.
