+++ compiler/sources/src/discovery/mod.rs
@@ -8,7 +8,6 @@
 //! The first match wins.
 
 use std::{
-    collections::HashSet,
     fs,
     path::{Path, PathBuf},
 };
@@ -16,10 +15,15 @@ use std::{
 use ironplc_dsl::core::FileId;
 use ironplc_dsl::diagnostic::{Diagnostic, Label};
 use ironplc_problems::Problem;
-use log::{info, trace};
+use log::info;
 
 use crate::file_type::FileType;
-use crate::libraries::{LibraryName, LibraryReference};
+use crate::libraries::LibraryReference;
+
+mod plcproj;
+mod sln;
+use plcproj::{collect_plcproj_via_walk, merge_plcproj_projects};
+use sln::resolve_plcproj_via_sln;
 
 /// The type of PLC project that was detected.
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -129,7 +133,7 @@ fn detect_beremiz(dir: &Path) -> Option<DiscoveredProject> {
 /// directory), which also rules out symlink cycles. Each directory's
 /// entries are sorted by name before recursing, so the result is
 /// deterministic regardless of filesystem iteration order.
-fn walk_files(dir: &Path, out: &mut Vec<PathBuf>) {
+pub(super) fn walk_files(dir: &Path, out: &mut Vec<PathBuf>) {
     let Ok(entries) = fs::read_dir(dir) else {
         return;
     };
@@ -183,412 +187,6 @@ fn detect_twincat(dir: &Path) -> Option<Result<DiscoveredProject, Diagnostic>> {
     }
 
     Some(merge_plcproj_projects(dir, plcproj_paths))
-}
-
-/// Find the single `.sln` file directly in `dir` (not recursive -- a
-/// `.sln` is always the file a user/tool points at directly, unlike
-/// `.plcproj`, which real layouts commonly nest several levels deep).
-///
-/// Returns `None` if there is no `.sln`, or more than one: ambiguous,
-/// callers should fall back to the recursive walk rather than guess.
-fn find_sln(dir: &Path) -> Option<PathBuf> {
-    let entries = fs::read_dir(dir).ok()?;
-
-    let mut sln_files: Vec<PathBuf> = entries
-        .filter_map(Result::ok)
-        .map(|entry| entry.path())
-        .filter(|path| {
-            path.is_file()
-                && path
-                    .extension()
-                    .is_some_and(|ext| ext.eq_ignore_ascii_case("sln"))
-        })
-        .collect();
-
-    if sln_files.len() != 1 {
-        return None;
-    }
-    sln_files.pop()
-}
-
-/// Resolve `.plcproj` paths via `.sln` -> `.tsproj` -> `PrjFilePath`.
-///
-/// Returns an empty vec if there's no unambiguous `.sln` in `dir`, or
-/// the `.sln` found doesn't lead to any `.plcproj` files -- callers
-/// should fall back to [`collect_plcproj_via_walk`] in that case.
-fn resolve_plcproj_via_sln(dir: &Path) -> Vec<PathBuf> {
-    let Some(sln_path) = find_sln(dir) else {
-        return Vec::new();
-    };
-
-    parse_sln(&sln_path)
-        .into_iter()
-        .flat_map(|tsproj_path| resolve_plcproj_via_tsproj(&tsproj_path))
-        .collect()
-}
-
-/// Parse a `.sln` (a line-oriented format, not XML) and return the
-/// `.tsproj` paths it lists, resolved relative to the `.sln`'s own
-/// directory.
-///
-/// Each project entry has the form:
-/// `Project("{TypeGUID}") = "Name", "RelativePath", "{ProjectGUID}"`.
-/// Entries whose `RelativePath` doesn't end in `.tsproj` are other
-/// Visual Studio project types the `.sln` may also list (a
-/// `DriveManager.tcdmproj`, a `Scope.tcmproj`, ...) and are ignored, not
-/// treated as an error.
-fn parse_sln(sln_path: &Path) -> Vec<PathBuf> {
-    let Ok(content) = fs::read_to_string(sln_path) else {
-        return Vec::new();
-    };
-    let sln_dir = sln_path.parent().unwrap_or(sln_path);
-
-    let mut tsproj_paths = Vec::new();
-    for line in content.lines() {
-        let line = line.trim();
-        if !line.starts_with("Project(") {
-            continue;
-        }
-
-        // Quoted segments land at odd indices after splitting on '"':
-        // [1]=TypeGUID, [3]=Name, [5]=RelativePath, [7]=ProjectGUID.
-        let quoted: Vec<&str> = line.split('"').collect();
-        let Some(relative_path) = quoted.get(5) else {
-            continue;
-        };
-
-        let is_tsproj = relative_path
-            .rsplit('.')
-            .next()
-            .is_some_and(|ext| ext.eq_ignore_ascii_case("tsproj"));
-        if !is_tsproj {
-            continue;
-        }
-
-        let normalized = relative_path.replace('\\', "/");
-        tsproj_paths.push(sln_dir.join(normalized));
-    }
-
-    tsproj_paths
-}
-
-/// Parse a `.tsproj` (XML) and return the `.plcproj` files it names via
-/// its nested `<Project PrjFilePath="...">` entries, resolved relative
-/// to the `.tsproj`'s own directory.
-///
-/// A `.tsproj` names a sub-project through `PrjFilePath` regardless of
-/// nesting depth, so entries are found by attribute presence
-/// (`.descendants()`, the same traversal `parse_plcproj` uses for
-/// `<Compile>`), not by assuming a fixed nesting shape. Entries whose
-/// `PrjFilePath` doesn't end in `.plcproj` -- notably `.splcproj`
-/// TwinSAFE safety projects, which use a different compilation model --
-/// are skipped, not treated as an error.
-///
-/// Returns an empty vec if the file can't be read or parsed as XML.
-fn resolve_plcproj_via_tsproj(tsproj_path: &Path) -> Vec<PathBuf> {
-    let Ok(content) = fs::read_to_string(tsproj_path) else {
-        return Vec::new();
-    };
-    let Ok(doc) = roxmltree::Document::parse(&content) else {
-        return Vec::new();
-    };
-    let tsproj_dir = tsproj_path.parent().unwrap_or(tsproj_path);
-
-    let mut plcproj_paths = Vec::new();
-    for node in doc.descendants() {
-        if !node.is_element() {
-            continue;
-        }
-        let Some(prj_file_path) = node.attribute("PrjFilePath") else {
-            continue;
-        };
-
-        let is_plcproj = prj_file_path
-            .rsplit('.')
-            .next()
-            .is_some_and(|ext| ext.eq_ignore_ascii_case("plcproj"));
-        if !is_plcproj {
-            continue;
-        }
-
-        let normalized = prj_file_path.replace('\\', "/");
-        plcproj_paths.push(tsproj_dir.join(normalized));
-    }
-
-    plcproj_paths
-}
-
-/// Recursively collect `.plcproj` candidates and keep only the first
-/// (sorted) per directory -- collapses same-directory duplicates (a
-/// previously observed stale-rename artifact) without discarding
-/// genuine sub-projects that live in different directories.
-///
-/// Searches recursively, since real TwinCAT layouts commonly nest
-/// `.plcproj` files several levels below the directory a user would
-/// naturally point the tool at.
-fn collect_plcproj_via_walk(dir: &Path) -> Vec<PathBuf> {
-    let mut files = Vec::new();
-    walk_files(dir, &mut files);
-
-    let mut candidates: Vec<PathBuf> = files
-        .into_iter()
-        .filter(|path| {
-            trace!("Check if file {path:?} is plcproj");
-            path.extension()
-                .is_some_and(|ext| ext.eq_ignore_ascii_case("plcproj"))
-        })
-        .collect();
-    candidates.sort();
-
-    let mut seen_dirs = HashSet::new();
-    let mut plcproj_paths: Vec<PathBuf> = Vec::new();
-    for path in candidates {
-        let dir_key = path.parent().unwrap_or(dir).to_path_buf();
-        if seen_dirs.insert(dir_key) {
-            plcproj_paths.push(path);
-        }
-    }
-    plcproj_paths
-}
-
-/// Parse and merge a resolved, non-empty list of `.plcproj` paths into
-/// one compilation unit.
-///
-/// A real solution commonly has more than one `.plcproj` -- a main PLC
-/// project plus one or more library/shared sub-projects that it (or
-/// each other) reference types from. All given `.plcproj` files are
-/// merged into a single compilation unit, the same principle already
-/// applied to LSP workspace folders.
-fn merge_plcproj_projects(
-    dir: &Path,
-    plcproj_paths: Vec<PathBuf>,
-) -> Result<DiscoveredProject, Diagnostic> {
-    // <Compile Include="..."> paths in a .plcproj are always relative to
-    // that .plcproj file's own directory, not the (possibly higher, now
-    // that files can be nested arbitrarily deep) directory originally
-    // passed to discover() -- each is parsed against its own directory
-    // regardless of how many sub-projects are being merged.
-    let single = plcproj_paths.len() == 1;
-    let mut merged_files = Vec::new();
-    let mut merged_errors = Vec::new();
-    let mut merged_library_references: Vec<LibraryReference> = Vec::new();
-    let mut seen_files = HashSet::new();
-    let mut seen_libraries: HashSet<LibraryName> = HashSet::new();
-    let mut merged_root_dir = dir.to_path_buf();
-
-    for plcproj_path in &plcproj_paths {
-        let plcproj_dir = plcproj_path.parent().unwrap_or(dir);
-        let project = parse_plcproj(plcproj_path, plcproj_dir)?;
-
-        if single {
-            merged_root_dir = project.root_dir.clone();
-        }
-
-        // A library referenced by more than one sub-project must only be
-        // activated once. Dedup by name (first reference wins), matching the
-        // name-only resolution the registry performs downstream.
-        for reference in project.library_references {
-            if seen_libraries.insert(reference.name.clone()) {
-                merged_library_references.push(reference);
-            }
-        }
-
-        for file in project.files {
-            // A file referenced by more than one sub-project (a shared
-            // dependency) must only be loaded/declared once. Dedup by
-            // canonical path, not the raw resolved path -- two
-            // sub-projects in different directories that both reach the
-            // same file via a relative `..` segment resolve to distinct,
-            // non-canonicalized paths that still name the same file.
-            let key = fs::canonicalize(&file).unwrap_or_else(|_| file.clone());
-            if seen_files.insert(key) {
-                merged_files.push(file);
-            }
-        }
-        merged_errors.extend(project.errors);
-    }
-
-    Ok(DiscoveredProject {
-        project_type: ProjectType::TwinCat,
-        root_dir: merged_root_dir,
-        files: merged_files,
-        library_references: merged_library_references,
-        errors: merged_errors,
-    })
-}
-
-/// Parse a `.plcproj` file and extract `<Compile Include="...">` paths.
-fn parse_plcproj(plcproj_path: &Path, root_dir: &Path) -> Result<DiscoveredProject, Diagnostic> {
-    let content = fs::read_to_string(plcproj_path).map_err(|e| {
-        Diagnostic::problem(
-            Problem::CannotReadFile,
-            Label::file(
-                FileId::from_path(plcproj_path),
-                format!("Cannot read .plcproj file: {e}"),
-            ),
-        )
-    })?;
-
-    let doc = roxmltree::Document::parse(&content).map_err(|e| {
-        Diagnostic::problem(
-            Problem::XmlMalformed,
-            Label::file(
-                FileId::from_path(plcproj_path),
-                format!("Malformed .plcproj XML: {e}"),
-            ),
-        )
-    })?;
-
-    let mut files = Vec::new();
-    let mut errors = Vec::new();
-    let library_references = parse_library_references(&doc, plcproj_path);
-
-    // Find all <Compile Include="..."> elements anywhere in the document.
-    // An entry that doesn't resolve to a real file (a stale reference, a
-    // case-sensitivity mismatch, a genuinely missing asset) is recorded
-    // as an error and skipped -- but does not abort the whole project:
-    // every other per-file problem in the codebase already works this
-    // way, and one bad reference shouldn't hide every other, perfectly
-    // valid file in the same project from ever being checked. The
-    // command as a whole must still fail, though (see `errors` field doc).
-    for node in doc.descendants() {
-        if node.is_element() && node.tag_name().name() == "Compile" {
-            if let Some(include) = node.attribute("Include") {
-                // Resolve relative to the .plcproj directory, normalizing
-                // Windows-style backslash separators
-                let normalized = include.replace('\\', "/");
-                let resolved = root_dir.join(&normalized);
-
-                if !resolved.is_file() {
-                    errors.push(Diagnostic::problem(
-                        Problem::CannotReadFile,
-                        Label::file(
-                            FileId::from_path(plcproj_path),
-                            format!(
-                                "Referenced file does not exist: {} (resolved to {})",
-                                include,
-                                resolved.display()
-                            ),
-                        ),
-                    ));
-                    continue;
-                }
-
-                files.push(resolved);
-            }
-        }
-    }
-
-    Ok(DiscoveredProject {
-        project_type: ProjectType::TwinCat,
-        root_dir: root_dir.to_path_buf(),
-        files,
-        library_references,
-        errors,
-    })
-}
-
-/// Extract the compatibility-library references from a parsed `.plcproj`.
-///
-/// A `.plcproj` states which libraries the project uses (the vendor's own
-/// record) in `<ItemGroup>` as one of two element types (`REQ-CL-sources-001`):
-///
-/// - `<PlaceholderReference Include="Name">` — the common, version-flexible
-///   form. The version lives in `<DefaultResolution>Name, Version (Vendor)`,
-///   usually the `*` wildcard.
-/// - `<LibraryReference Include="Name,Version,Vendor">` — a concrete, pinned
-///   reference.
-///
-/// Both may carry a `<Namespace>` child (the qualifier the source may write).
-/// A reference marked `<SystemLibrary>true</SystemLibrary>` (CODESYS /
-/// visualization system libraries such as `VisuElems`) is not vendor-authored
-/// ST we bundle, so the first increment skips it. Element names are matched by
-/// local name, so the MSBuild default `xmlns` does not need special handling —
-/// the same way `<Compile>` is already read.
-fn parse_library_references(
-    doc: &roxmltree::Document,
-    plcproj_path: &Path,
-) -> Vec<LibraryReference> {
-    let declared_in = FileId::from_path(plcproj_path);
-    let mut references = Vec::new();
-
-    for node in doc.descendants() {
-        if !node.is_element() {
-            continue;
-        }
-
-        let (raw_name, version) = match node.tag_name().name() {
-            "PlaceholderReference" => {
-                let Some(include) = node.attribute("Include") else {
-                    continue;
-                };
-                // The version, when present, is the middle field of
-                // `<DefaultResolution>Name, Version (Vendor)</DefaultResolution>`.
-                let version = child_text(node, "DefaultResolution")
-                    .and_then(|resolution| default_resolution_version(&resolution));
-                (include.to_string(), version)
-            }
-            "LibraryReference" => {
-                let Some(include) = node.attribute("Include") else {
-                    continue;
-                };
-                // Include is `Name,Version,Vendor`.
-                let mut fields = include.splitn(3, ',');
-                let name = fields.next().unwrap_or("").trim().to_string();
-                let version = fields
-                    .next()
-                    .map(|field| field.trim().to_string())
-                    .filter(|field| !field.is_empty());
-                (name, version)
-            }
-            _ => continue,
-        };
-
-        // Skip system libraries for now (REQ-CL-sources-001 note).
-        if child_text(node, "SystemLibrary")
-            .is_some_and(|value| value.trim().eq_ignore_ascii_case("true"))
-        {
-            continue;
-        }
-
-        let name = raw_name.trim();
-        if name.is_empty() {
-            continue;
-        }
-
-        let namespace = child_text(node, "Namespace")
-            .map(|value| value.trim().to_string())
-            .filter(|value| !value.is_empty());
-
-        references.push(LibraryReference {
-            name: LibraryName::from(name),
-            version,
-            namespace,
-            declared_in: declared_in.clone(),
-        });
-    }
-
-    references
-}
-
-/// The trimmed text of the first direct child element named `tag`, if any.
-fn child_text(node: roxmltree::Node, tag: &str) -> Option<String> {
-    node.children()
-        .find(|child| child.is_element() && child.tag_name().name() == tag)
-        .and_then(|child| child.text())
-        .map(str::to_string)
-}
-
-/// Extract the version from a `<DefaultResolution>` value of the form
-/// `Name, Version (Vendor)`. Returns `None` when no version is present.
-fn default_resolution_version(resolution: &str) -> Option<String> {
-    let after_name = resolution.split_once(',')?.1;
-    let version = after_name.split('(').next()?.trim();
-    if version.is_empty() {
-        None
-    } else {
-        Some(version.to_string())
-    }
 }
 
 /// Fallback detection: recursively enumerate all supported files under
@@ -1348,250 +946,5 @@ mod tests {
         // Only the real file is found; the symlink is not traversed.
         assert_eq!(result.files.len(), 1);
         assert!(result.files[0].ends_with("real/main.st"));
-    }
-
-    // -- .sln -> .tsproj -> PrjFilePath discovery tests --
-
-    fn write_sln(dir: &Path, name: &str, tsproj_entries: &[(&str, &str)]) {
-        let mut content = String::from(
-            "Microsoft Visual Studio Solution File, Format Version 12.00\n\
-             # TcXaeShell Solution File, Format Version 11.00\n",
-        );
-        for (project_name, relative_path) in tsproj_entries {
-            content.push_str(&format!(
-                "Project(\"{{B1E792BE-AA5F-4E3C-8C82-674BF9C0715B}}\") = \"{project_name}\", \"{relative_path}\", \"{{9406D69C-EBA9-4591-A513-578A75D14426}}\"\nEndProject\n"
-            ));
-        }
-        fs::write(dir.join(name), content).unwrap();
-    }
-
-    fn write_tsproj(path: &Path, plc_entries: &[(&str, &str)]) {
-        let mut inner = String::new();
-        for (name, prj_file_path) in plc_entries {
-            inner.push_str(&format!(
-                r#"<Project GUID="{{6DADE760-7FAC-4830-92BA-478C8595D673}}" Name="{name}" PrjFilePath="{prj_file_path}" />"#
-            ));
-        }
-        let content = format!(
-            r#"<Project ProjectGUID="{{9406D69C-EBA9-4591-A513-578A75D14426}}">{inner}</Project>"#
-        );
-        fs::write(path, content).unwrap();
-    }
-
-    #[test]
-    fn discover_when_sln_present_then_resolves_via_tsproj() {
-        let dir = TempDir::new().unwrap();
-        write_sln(dir.path(), "Solution.sln", &[("Main", "Main\\Main.tsproj")]);
-
-        let tsproj_dir = dir.path().join("Main");
-        fs::create_dir_all(&tsproj_dir).unwrap();
-        write_tsproj(
-            &tsproj_dir.join("Main.tsproj"),
-            &[("MainRuntime", "MainRuntime\\MainRuntime.plcproj")],
-        );
-
-        let plcproj_dir = tsproj_dir.join("MainRuntime");
-        fs::create_dir_all(&plcproj_dir).unwrap();
-        fs::write(plcproj_dir.join("A.TcPOU"), "<TcPlcObject/>").unwrap();
-        fs::write(
-            plcproj_dir.join("MainRuntime.plcproj"),
-            r#"<Project>
-  <ItemGroup>
-    <Compile Include="A.TcPOU" />
-  </ItemGroup>
-</Project>"#,
-        )
-        .unwrap();
-
-        let result = discover(dir.path()).unwrap();
-
-        assert_eq!(result.project_type, ProjectType::TwinCat);
-        assert_eq!(result.files.len(), 1);
-        assert!(result.files[0].ends_with("A.TcPOU"));
-    }
-
-    #[test]
-    fn discover_when_sln_lists_non_tsproj_project_then_ignored() {
-        let dir = TempDir::new().unwrap();
-        write_sln(
-            dir.path(),
-            "Solution.sln",
-            &[
-                ("Scope", "Scope\\Scope.tcmproj"),
-                ("Main", "Main\\Main.tsproj"),
-            ],
-        );
-
-        let tsproj_dir = dir.path().join("Main");
-        fs::create_dir_all(&tsproj_dir).unwrap();
-        write_tsproj(
-            &tsproj_dir.join("Main.tsproj"),
-            &[("MainRuntime", "MainRuntime\\MainRuntime.plcproj")],
-        );
-
-        let plcproj_dir = tsproj_dir.join("MainRuntime");
-        fs::create_dir_all(&plcproj_dir).unwrap();
-        fs::write(plcproj_dir.join("A.TcPOU"), "<TcPlcObject/>").unwrap();
-        fs::write(
-            plcproj_dir.join("MainRuntime.plcproj"),
-            r#"<Project>
-  <ItemGroup>
-    <Compile Include="A.TcPOU" />
-  </ItemGroup>
-</Project>"#,
-        )
-        .unwrap();
-        // Scope.tcmproj deliberately doesn't exist -- a non-.tsproj entry
-        // must not be resolved/read at all, only filtered out by extension.
-
-        let result = discover(dir.path()).unwrap();
-
-        assert_eq!(result.files.len(), 1);
-        assert!(result.files[0].ends_with("A.TcPOU"));
-    }
-
-    #[test]
-    fn discover_when_sln_and_stale_duplicate_plcproj_then_picks_named_one() {
-        // Regression case modeled on a real private-corpus TwinCAT
-        // solution: a project renamed Foo -> Fooo left a stale
-        // `Foo.plcproj` and `Foo.tsproj` behind, neither referenced by
-        // the .sln. The stale file is named so it would sort first,
-        // so this fails under plain alphabetical-glob resolution.
-        let dir = TempDir::new().unwrap();
-        write_sln(dir.path(), "Solution.sln", &[("Fooo", "Fooo.tsproj")]);
-
-        // Live .tsproj, referenced by the .sln.
-        write_tsproj(
-            &dir.path().join("Fooo.tsproj"),
-            &[("FoooRuntime", "Runtime\\Fooo.plcproj")],
-        );
-        // Stale .tsproj, NOT referenced by the .sln -- must be ignored.
-        write_tsproj(
-            &dir.path().join("Foo.tsproj"),
-            &[("FooRuntime", "Runtime\\Foo.plcproj")],
-        );
-
-        let plcproj_dir = dir.path().join("Runtime");
-        fs::create_dir_all(&plcproj_dir).unwrap();
-        fs::write(plcproj_dir.join("LIVE.TcPOU"), "<TcPlcObject/>").unwrap();
-        fs::write(
-            plcproj_dir.join("Fooo.plcproj"),
-            r#"<Project>
-  <ItemGroup>
-    <Compile Include="LIVE.TcPOU" />
-  </ItemGroup>
-</Project>"#,
-        )
-        .unwrap();
-        // Stale duplicate: sorts before "Fooo.plcproj" alphabetically,
-        // so the old glob-only logic would have picked this one.
-        fs::write(plcproj_dir.join("Foo.plcproj"), "<Project/>").unwrap();
-
-        let result = discover(dir.path()).unwrap();
-
-        assert_eq!(result.files.len(), 1);
-        assert!(result.files[0].ends_with("LIVE.TcPOU"));
-    }
-
-    #[test]
-    fn discover_when_tsproj_references_splcproj_then_skipped() {
-        let dir = TempDir::new().unwrap();
-        write_sln(dir.path(), "Solution.sln", &[("Main", "Main.tsproj")]);
-        write_tsproj(
-            &dir.path().join("Main.tsproj"),
-            &[
-                ("MainRuntime", "MainRuntime\\MainRuntime.plcproj"),
-                ("MainTwinSAFE", "MainTwinSAFE\\MainTwinSAFE.splcproj"),
-            ],
-        );
-
-        let plcproj_dir = dir.path().join("MainRuntime");
-        fs::create_dir_all(&plcproj_dir).unwrap();
-        fs::write(plcproj_dir.join("A.TcPOU"), "<TcPlcObject/>").unwrap();
-        fs::write(
-            plcproj_dir.join("MainRuntime.plcproj"),
-            r#"<Project>
-  <ItemGroup>
-    <Compile Include="A.TcPOU" />
-  </ItemGroup>
-</Project>"#,
-        )
-        .unwrap();
-        // MainTwinSAFE.splcproj deliberately doesn't exist -- proves it
-        // was never resolved/read, only filtered out by extension.
-
-        let result = discover(dir.path()).unwrap();
-
-        assert_eq!(result.files.len(), 1);
-        assert!(result.files[0].ends_with("A.TcPOU"));
-    }
-
-    #[test]
-    fn discover_when_multiple_sln_at_top_level_then_falls_back_to_walk() {
-        let dir = TempDir::new().unwrap();
-        write_sln(dir.path(), "A.sln", &[]);
-        write_sln(dir.path(), "B.sln", &[]);
-
-        fs::write(dir.path().join("MAIN.TcPOU"), "<TcPlcObject/>").unwrap();
-        fs::write(
-            dir.path().join("project.plcproj"),
-            r#"<Project>
-  <ItemGroup>
-    <Compile Include="MAIN.TcPOU" />
-  </ItemGroup>
-</Project>"#,
-        )
-        .unwrap();
-
-        let result = discover(dir.path()).unwrap();
-
-        assert_eq!(result.project_type, ProjectType::TwinCat);
-        assert_eq!(result.files.len(), 1);
-        assert!(result.files[0].ends_with("MAIN.TcPOU"));
-    }
-
-    #[test]
-    fn discover_when_sln_references_tsproj_with_multiple_plcproj_then_merges_all() {
-        let dir = TempDir::new().unwrap();
-        write_sln(dir.path(), "Solution.sln", &[("Main", "Main.tsproj")]);
-        write_tsproj(
-            &dir.path().join("Main.tsproj"),
-            &[
-                ("MainRuntime", "MainRuntime\\MainRuntime.plcproj"),
-                ("SharedLib", "SharedLib\\SharedLib.plcproj"),
-            ],
-        );
-
-        let main_dir = dir.path().join("MainRuntime");
-        fs::create_dir_all(&main_dir).unwrap();
-        fs::write(main_dir.join("MAIN.TcPOU"), "<TcPlcObject/>").unwrap();
-        fs::write(
-            main_dir.join("MainRuntime.plcproj"),
-            r#"<Project>
-  <ItemGroup>
-    <Compile Include="MAIN.TcPOU" />
-  </ItemGroup>
-</Project>"#,
-        )
-        .unwrap();
-
-        let lib_dir = dir.path().join("SharedLib");
-        fs::create_dir_all(&lib_dir).unwrap();
-        fs::write(lib_dir.join("FB_Shared.TcPOU"), "<TcPlcObject/>").unwrap();
-        fs::write(
-            lib_dir.join("SharedLib.plcproj"),
-            r#"<Project>
-  <ItemGroup>
-    <Compile Include="FB_Shared.TcPOU" />
-  </ItemGroup>
-</Project>"#,
-        )
-        .unwrap();
-
-        let result = discover(dir.path()).unwrap();
-
-        assert_eq!(result.files.len(), 2);
-        assert!(result.files.iter().any(|f| f.ends_with("MAIN.TcPOU")));
-        assert!(result.files.iter().any(|f| f.ends_with("FB_Shared.TcPOU")));
     }
 }
